Добавлены поля путь и имя изображения товара в базу, тип товара стал int, добавлена обработка обращения к серверу с выдачей из базы товара по типу и всех для конкретной POS

+++ src/com/pos/Service/Dto/CheckDto.java
@@ -53,7 +53,7 @@ public class CheckDto {
         goodsDtoList.clear();
     }
 
-    public void addGoods(Long i) {
+    public void addGoods(int i) {
 
         Goods currGoods = new Goods();
         currGoods.setGoodsType(i);
